@@ -4,7 +4,7 @@ buildscript {
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath 'com.android.tools.build:gradle:3.4.1'
         classpath 'com.novoda:bintray-release:0.8.0'
     }
 }
@@ -29,9 +29,14 @@ ext {
     publishVersion = '1.4.9'
     licences = ['Apache-2.0']
 
-    compileSdkVersion = 27
-    buildToolsVersion = '26.0.2'
-    targetSdkVersion = 27
+    compileSdkVersion = 28
+    buildToolsVersion = '28.0.0'
+    targetSdkVersion = 28
 
-    supportLibVersion = '27.1.1'
+    appcompatVersion = '1.0.2'
+    recyclerViewVersion = '1.0.0'
+    cardViewVersion = '1.0.0'
+    materialVersion = '1.0.0'
+
+    glideVersion = '4.9.0'
 }
